+++ src/isovec/__init__.py
@@ -1,8 +1,9 @@
 
-from .isotope  import Isotope
-from .element  import Element
-from .molecule import Molecule
-from .mixture  import Mixture
+from .isotope   import Isotope
+from .substance import Substance
+from .element   import Element
+from .molecule  import Molecule
+from .mixture   import Mixture
 
 from .isotopes import *
 from .elements import *