+++ src/isovec/element.py
@@ -4,6 +4,9 @@ from collections import defaultdict
 
 from .substance import Substance
 from .isotope import Isotope
+from .isotopes import _natural_compositions
+
+from .constants import ATOM_NUMB_TO_SYMBOL
 
 
 class Element(Substance):
@@ -15,14 +18,20 @@ class Element(Substance):
         """Returns a tuple of allowed classes for constituents."""
         return (Isotope,)
 
-    def __init__(self, name: str, isotopes: dict[Isotope, float], mode: str = "_legacy", **kwargs) -> None:
+    def __init__(self, name: str, isotopes: dict[Isotope, float], mode: str = "_legacy", natural: bool = False, **kwargs) -> None:
 
-        super().__init__(name=name, constituents=isotopes, mode=mode, kwargs=kwargs)
+        super().__init__(name=name, constituents=isotopes, mode=mode, **kwargs)
+
+        self._is_natural = natural
 
         # take atomic number of first isotope (and make sure all isotopes are of the same element)
         self._Z: int = next(iter(isotopes)).Z  # atomic number of element
         if not all(self._Z == isotope.Z for isotope in isotopes.keys()):
             raise ValueError(f"Atomic number of all isotopes of {self.__class__.__name__} \"{self._name}\" must match!")
+        
+        # search for symbol
+        if self._symbol == "":
+            self._symbol = ATOM_NUMB_TO_SYMBOL[self._Z]
 
     # override
     @classmethod
@@ -50,7 +59,7 @@ class Element(Substance):
     # ########
 
     def _calc_A_r(self) -> float:
-        """Calculates relative atomic mass (atomic weight) [-] of the element.
+        r"""Calculates relative atomic mass (atomic weight) [-] of the element.
         
         The relative atomic masses of all isotopes are weighted by their atomic fraction and summed up:
             $$\overline{M} = \sum_i \left( x_i \cdot A_\mathrm{r},i \right)$$
@@ -63,9 +72,18 @@ class Element(Substance):
     # ########
 
     # override
-    def _append_isotopes(self, dict_list: defaultdict[Isotope, list[float]], x_p: float = 1.0) -> defaultdict[Isotope, list[float]]:
-        for isotope, x_i in self._constituents.items():
-            dict_list[isotope].append(x_p*x_i)
+    def _append_isotopes(self, dict_list: defaultdict[Isotope, list[float]], by_weight: bool = False, f_p: float = 1.0, use_natural: bool = False) -> defaultdict[Isotope, list[float]]:
+        if not by_weight:
+            constituents = self._constituents
+        else:
+            constituents = self.get_constituents_in_wt() 
+        if use_natural and self._is_natural:
+            natural_isotope = _natural_compositions[self._Z]
+            f_i = sum(constituents.values())
+            dict_list[natural_isotope].append(f_p*f_i)
+        else:
+            for isotope, f_i in constituents.items():
+                dict_list[isotope].append(f_p*f_i)
         return dict_list